LPS-57357 Move Control Panel Entry class
